+++ R/main_bootstrap.R
@@ -8,14 +8,16 @@
 #' @description
 #' A function for bootstrapping coefficients of generalized autoregressive score (GAS) models of Creal et al. (2013) and Harvey (2013).
 #' Method \code{"parametric"} repeatedly simulates time series using the parametric model and re-estimates coefficients.
+#' Methods \code{"simple_block"}, \code{"moving_block"}, and \code{"stationary_block"} perform the standard variations of the circular block bootstrap.
 #' Instead of supplying arguments about the model, the function can be applied to the \code{gas} object obtained by the \code{\link[gasmodel:gas]{gas()}} function.
 #' The function enables parallelization.
 #'
 #' @inheritParams gas
 #' @inheritParams gas_simulate
 #'
-#' @param method A method used for bootstrapping. Currently, the only supported method is \code{"parametric"}.
+#' @param method A method used for bootstrapping. Supported methods are \code{"parametric"}, \code{"simple_block"}, \code{"moving_block"}, and \code{"stationary_block"}.
 #' @param rep_boot A number of bootstrapping repetitions.
+#' @param block_length A length of blocks for methods \code{"simple_block"} and \code{"moving_block"}. A mean length of blocks for method \code{"stationary_block"}.
 #' @param quant A numeric vector of probabilities determining quantiles.
 #' @param parallel_function A parallelization function. For suitable wrappers of common R parallelization functions, see \code{\link[gasmodel:wrappers_parallel]{wrappers_parallel}}. Can be set to \code{NULL} if no parallelization is to be used.
 #' @param parallel_arguments An optional list of arguments to be passed to the optimization function.
@@ -78,9 +80,9 @@
 #' boot_gas}
 #'
 #' @export
-gas_bootstrap <- function(gas_object = NULL, method = "parametric", rep_boot = 1000L, quant = c(0.025, 0.975), y = NULL, x = NULL, distr = NULL, param = NULL, scaling = "unit", regress = "joint", p = 1L, q = 1L, par_static = NULL, par_link = NULL, par_init = NULL, lik_skip = 0L, coef_fix_value = NULL, coef_fix_other = NULL, coef_fix_special = NULL, coef_bound_lower = NULL, coef_bound_upper = NULL, coef_est = NULL, optim_function = wrapper_optim_nloptr, optim_arguments = list(opts = list(algorithm = 'NLOPT_LN_NELDERMEAD', xtol_rel = 0, maxeval = 1e6)), parallel_function = NULL, parallel_arguments = list()) {
+gas_bootstrap <- function(gas_object = NULL, method = "parametric", rep_boot = 1000L, block_length = NULL, quant = c(0.025, 0.975), y = NULL, x = NULL, distr = NULL, param = NULL, scaling = "unit", regress = "joint", p = 1L, q = 1L, par_static = NULL, par_link = NULL, par_init = NULL, lik_skip = 0L, coef_fix_value = NULL, coef_fix_other = NULL, coef_fix_special = NULL, coef_bound_lower = NULL, coef_bound_upper = NULL, coef_est = NULL, optim_function = wrapper_optim_nloptr, optim_arguments = list(opts = list(algorithm = 'NLOPT_LN_NELDERMEAD', xtol_rel = 0, maxeval = 1e6)), parallel_function = NULL, parallel_arguments = list()) {
   if (!is.null(gas_object) && inherits(gas_object, "gas")) {
-    gas_bootstrap(gas_object = NULL, method = method, rep_boot = rep_boot, quant = quant, y = gas_object$data$y, x = gas_object$data$x, distr = gas_object$model$distr, param = gas_object$model$param, scaling = gas_object$model$scaling, regress = gas_object$model$regress, p = gas_object$model$p, q = gas_object$model$q, par_static = gas_object$model$par_static, par_link = gas_object$model$par_link, par_init = gas_object$model$par_init, lik_skip = gas_object$model$lik_skip, coef_fix_value = gas_object$model$coef_fix_value, coef_fix_other = gas_object$model$coef_fix_other, coef_fix_special = gas_object$model$coef_fix_special, coef_bound_lower = gas_object$model$coef_bound_lower, coef_bound_upper = gas_object$model$coef_bound_upper, coef_est = gas_object$fit$coef_est, optim_function = gas_object$control$optim_function, optim_arguments = gas_object$control$optim_arguments, parallel_function = parallel_function, parallel_arguments = parallel_arguments)
+    gas_bootstrap(gas_object = NULL, method = method, rep_boot = rep_boot, block_length = block_length, quant = quant, y = gas_object$data$y, x = gas_object$data$x, distr = gas_object$model$distr, param = gas_object$model$param, scaling = gas_object$model$scaling, regress = gas_object$model$regress, p = gas_object$model$p, q = gas_object$model$q, par_static = gas_object$model$par_static, par_link = gas_object$model$par_link, par_init = gas_object$model$par_init, lik_skip = gas_object$model$lik_skip, coef_fix_value = gas_object$model$coef_fix_value, coef_fix_other = gas_object$model$coef_fix_other, coef_fix_special = gas_object$model$coef_fix_special, coef_bound_lower = gas_object$model$coef_bound_lower, coef_bound_upper = gas_object$model$coef_bound_upper, coef_est = gas_object$fit$coef_est, optim_function = gas_object$control$optim_function, optim_arguments = gas_object$control$optim_arguments, parallel_function = parallel_function, parallel_arguments = parallel_arguments)
   } else if (!is.null(gas_object)) {
     stop("Unsupported class of gas_object.")
   } else {
@@ -138,7 +140,7 @@ gas_bootstrap <- function(gas_object = NULL, method = "parametric", rep_boot = 1
       control$parallel_arguments <- check_generic_list(arg = parallel_arguments, arg_name = "parallel_arguments")
     }
     bootstrap <- list()
-    bootstrap$method <- check_my_method(method = method, values = c("parametric"))
+    bootstrap$method <- check_my_method(method = method, values = c("parametric", "simple_block", "moving_block", "stationary_block"))
     comp <- list()
     comp$rep_boot <- check_generic_positive_integer_scalar(arg = rep_boot, arg_name = "rep_boot")
     bootstrap$coef_set <- name_matrix(matrix(NA_real_, nrow = comp$rep_boot, ncol = info_coef$coef_num), paste0("coef", 1:comp$rep_boot), info_coef$coef_names)
@@ -147,132 +149,248 @@ gas_bootstrap <- function(gas_object = NULL, method = "parametric", rep_boot = 1
     comp$theta_bound_lower <- convert_coef_vector_to_theta_vector(model$coef_bound_lower, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
     comp$theta_bound_upper <- convert_coef_vector_to_theta_vector(model$coef_bound_upper, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
     comp$est_details <- list(data = data, model = model, fun = fun, info_distr = info_distr, info_par = info_par, info_coef = info_coef, print_progress = FALSE)
-    if (bootstrap$method == "parametric") {
-      comp$pre_num <- max(c(model$p, model$q, 1L))
-      comp$burn_num <- 10L + max(model$p) + max(model$q)
-      comp$full_num <- comp$pre_num + comp$burn_num + model$t
-      comp$average_x <- lapply(1:info_par$par_num, function(i) { colMeans(data$x[[i]], na.rm = TRUE) })
-      comp$y <- matrix(NA_real_, nrow = comp$full_num, ncol = model$n)
-      comp$x <- lapply(1:info_par$par_num, function(i) { rbind(matrix(NA_real_, nrow = comp$pre_num, ncol = model$m[i]), matrix(comp$average_x[[i]], nrow = comp$burn_num, ncol = model$m[i], byrow = TRUE), data$x[[i]]) })
-      comp$par_tv <- matrix(NA_real_, nrow = comp$full_num, ncol = info_par$par_num)
-      comp$score_tv <- matrix(NA_real_, nrow = comp$full_num, ncol = info_par$par_num)
-      comp$idx_na <- which(rowSums(sapply(comp$x, function(e) { rowSums(is.na(e)) })) > 0L)
-      comp$idx_ok <- which(!((1:comp$full_num) %in% comp$idx_na))
-      comp$struc <- convert_coef_vector_to_struc_list(coef_vec = model$coef_est, m = model$m, p = model$p, q = model$q, par_names = info_par$par_names, par_of_coef_names = info_coef$par_of_coef_names)
-      comp$omega_vector <- sapply(comp$struc, function(e) { e$omega })
-      comp$beta_list <- lapply(comp$struc, function(e) { e$beta })
-      comp$alpha_list <- lapply(comp$struc, function(e) { e$alpha })
-      comp$phi_list <- lapply(comp$struc, function(e) { e$phi })
-      if (all(model$p + model$q == 0L)) {
-        comp$par_init <- model$par_init
-        if (any(is.na(comp$par_init))) {
-          comp$par_unc <- sapply(1:info_par$par_num, function(i) { (comp$omega_vector[i] + comp$average_x[[i]] %*% comp$beta_list[[i]]) })
-          comp$par_init[is.na(comp$par_init)] <- comp$par_unc[is.na(comp$par_init)]
+    if (bootstrap$method == "parametric" && all(model$p + model$q == 0L)) {
+      comp$run_details <- list(data = data, model = model, control = control, info_distr = info_distr, info_par = info_par, info_coef = info_coef, comp = comp)
+      comp$run_fun <- function(id, run_details) {
+        data <- run_details$data
+        model <- run_details$model
+        control <- run_details$control
+        info_distr <- run_details$info_distr
+        info_par <- run_details$info_par
+        info_coef <- run_details$info_coef
+        comp <- run_details$comp
+        pre_num <- max(c(model$p, model$q, 1L))
+        burn_num <- 10L + max(model$p) + max(model$q)
+        full_num <- pre_num + burn_num + model$t
+        average_x <- lapply(1:info_par$par_num, function(i) { colMeans(data$x[[i]], na.rm = TRUE) })
+        y <- matrix(NA_real_, nrow = full_num, ncol = model$n)
+        x <- lapply(1:info_par$par_num, function(i) { rbind(matrix(NA_real_, nrow = pre_num, ncol = model$m[i]), matrix(average_x[[i]], nrow = burn_num, ncol = model$m[i], byrow = TRUE), data$x[[i]]) })
+        par_tv <- matrix(NA_real_, nrow = full_num, ncol = info_par$par_num)
+        score_tv <- matrix(NA_real_, nrow = full_num, ncol = info_par$par_num)
+        idx_na <- which(rowSums(sapply(x, function(e) { rowSums(is.na(e)) })) > 0L)
+        idx_ok <- which(!((1:full_num) %in% idx_na))
+        struc <- convert_coef_vector_to_struc_list(coef_vec = model$coef_est, m = model$m, p = model$p, q = model$q, par_names = info_par$par_names, par_of_coef_names = info_coef$par_of_coef_names)
+        omega_vector <- sapply(struc, function(e) { e$omega })
+        beta_list <- lapply(struc, function(e) { e$beta })
+        alpha_list <- lapply(struc, function(e) { e$alpha })
+        phi_list <- lapply(struc, function(e) { e$phi })
+        par_init <- model$par_init
+        if (any(is.na(par_init))) {
+          par_unc <- sapply(1:info_par$par_num, function(i) { (omega_vector[i] + average_x[[i]] %*% beta_list[[i]]) })
+          par_init[is.na(par_init)] <- par_unc[is.na(par_init)]
         }
-        if (length(comp$idx_na) > 0L) {
-          comp$par_tv[comp$idx_na, ] <- matrix(comp$par_init, nrow = length(comp$idx_na), ncol = info_par$par_num, byrow = TRUE)
-          comp$score_tv[comp$idx_na, ] <- 0
+        if (length(idx_na) > 0L) {
+          par_tv[idx_na, ] <- matrix(par_init, nrow = length(idx_na), ncol = info_par$par_num, byrow = TRUE)
+          score_tv[idx_na, ] <- 0
         }
-        comp$run_details <- list(model = model, control = control, info_par = info_par, comp = comp)
-        comp$run_fun <- function(id, run_details) {
-          model <- run_details$model
-          control <- run_details$control
-          info_par <- run_details$info_par
-          comp <- run_details$comp
-          comp$par_tv[comp$idx_ok, ] <- matrix(comp$omega_vector, nrow = length(comp$idx_ok), ncol = info_par$par_num, byrow = TRUE)
-          if (any(model$m > 0L)) {
-            comp$par_tv[comp$idx_ok, ] <- comp$par_tv[comp$idx_ok, ] + sapply(1L:info_par$par_num, function(i) { comp$x[[i]][comp$idx_ok, , drop = FALSE] %*% comp$beta_list[[i]] })
-          }
-          for (j in comp$idx_ok) {
-            comp$y[j, ] <- fun$random(t = 1L, f = comp$par_tv[j, , drop = FALSE])
-            comp$score_tv[j, ] <- fun$score(y = comp$y[j, , drop = FALSE], f = comp$par_tv[j, , drop = FALSE])
-          }
-          comp$est_details$data$y <- comp$y[(comp$pre_num + comp$burn_num + 1L):comp$full_num, , drop = FALSE]
-          comp$result_optim <- do.call(control$optim_function, args = c(list(obj_fun = comp$fun$likelihood_objective, theta_start = comp$theta_start, theta_bound_lower = comp$theta_bound_lower, theta_bound_upper = comp$theta_bound_upper, est_details = comp$est_details), control$optim_arguments))
-          coef_set <- comp$fun$convert_theta_vector_to_coef_vector(comp$result_optim$theta_optim, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
-          return(coef_set)
+        par_tv[idx_ok, ] <- matrix(omega_vector, nrow = length(idx_ok), ncol = info_par$par_num, byrow = TRUE)
+        if (any(model$m > 0L)) {
+          par_tv[idx_ok, ] <- par_tv[idx_ok, ] + sapply(1L:info_par$par_num, function(i) { x[[i]][idx_ok, , drop = FALSE] %*% beta_list[[i]] })
         }
-      } else if (model$regress == "joint") {
-        comp$par_init <- model$par_init
-        if (any(is.na(comp$par_init))) {
-          comp$par_unc <- sapply(1:info_par$par_num, function(i) { (comp$omega_vector[i] + comp$average_x[[i]] %*% comp$beta_list[[i]]) / (1 - sum(comp$phi_list[[i]])) })
-          comp$par_init[is.na(comp$par_init)] <- comp$par_unc[is.na(comp$par_init)]
+        for (j in idx_ok) {
+          y[j, ] <- fun$random(t = 1L, f = par_tv[j, , drop = FALSE])
+          score_tv[j, ] <- fun$score(y = y[j, , drop = FALSE], f = par_tv[j, , drop = FALSE])
         }
-        if (length(comp$idx_na) > 0L) {
-          comp$par_tv[comp$idx_na, ] <- matrix(comp$par_init, nrow = length(comp$idx_na), ncol = info_par$par_num, byrow = TRUE)
-          comp$score_tv[comp$idx_na, ] <- 0
-        }
-        comp$run_details <- list(model = model, control = control, info_par = info_par, comp = comp)
-        comp$run_fun <- function(id, run_details) {
-          model <- run_details$model
-          control <- run_details$control
-          info_par <- run_details$info_par
-          comp <- run_details$comp
-          comp$par_tv[comp$idx_ok, ] <- matrix(comp$omega_vector, nrow = length(comp$idx_ok), ncol = info_par$par_num, byrow = TRUE)
-          if (any(model$m > 0L)) {
-            comp$par_tv[comp$idx_ok, ] <- comp$par_tv[comp$idx_ok, ] + sapply(1L:info_par$par_num, function(i) { comp$x[[i]][comp$idx_ok, , drop = FALSE] %*% comp$beta_list[[i]] })
-          }
-          cur_e <- rep(NA_real_, info_par$par_num)
-          for (j in comp$idx_ok) {
-            for (i in 1:info_par$par_num) {
-              cur_e[i] <- sum(comp$par_tv[j - seq_along(comp$phi_list[[i]]), i] * comp$phi_list[[i]]) + sum(comp$score_tv[j - seq_along(comp$alpha_list[[i]]), i] * comp$alpha_list[[i]])
-            }
-            comp$par_tv[j, ] <- comp$par_tv[j, ] + cur_e
-            comp$y[j, ] <- fun$random(t = 1L, f = comp$par_tv[j, , drop = FALSE])
-            comp$score_tv[j, ] <- fun$score(y = comp$y[j, , drop = FALSE], f = comp$par_tv[j, , drop = FALSE])
-          }
-          comp$est_details$data$y <- comp$y[(comp$pre_num + comp$burn_num + 1L):comp$full_num, , drop = FALSE]
-          comp$result_optim <- do.call(control$optim_function, args = c(list(obj_fun = likelihood_objective, theta_start = comp$theta_start, theta_bound_lower = comp$theta_bound_lower, theta_bound_upper = comp$theta_bound_upper, est_details = comp$est_details), control$optim_arguments))
-          coef_set <- convert_theta_vector_to_coef_vector(comp$result_optim$theta_optim, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
-          return(coef_set)
-        }
-      } else if (model$regress == "sep") {
-        comp$err_tv <- matrix(NA_real_, nrow = comp$full_num, ncol = info_par$par_num)
-        comp$err_init <- model$par_init
-        comp$par_init <- model$par_init
-        if (any(is.na(model$par_init))) {
-          comp$par_unc <- sapply(1:info_par$par_num, function(i) { (comp$omega_vector[i] + comp$average_x[[i]] %*% comp$beta_list[[i]]) })
-          comp$par_init[is.na(comp$par_init)] <- comp$par_unc[is.na(comp$par_init)]
-          comp$err_init[is.na(comp$err_init)] <- 0
-        }
-        if (length(comp$idx_na) > 0L) {
-          comp$err_tv[comp$idx_na, ] <- matrix(comp$err_init, nrow = length(comp$idx_na), ncol = info_par$par_num, byrow = TRUE)
-          comp$par_tv[comp$idx_na, ] <- matrix(comp$par_init, nrow = length(comp$idx_na), ncol = info_par$par_num, byrow = TRUE)
-          comp$score_tv[comp$idx_na, ] <- 0
-        }
-        comp$run_details <- list(model = model, control = control, info_par = info_par, comp = comp)
-        comp$run_fun <- function(id, run_details) {
-          model <- run_details$model
-          control <- run_details$control
-          info_par <- run_details$info_par
-          comp <- run_details$comp
-          comp$par_tv[comp$idx_ok, ] <- matrix(comp$omega_vector, nrow = length(comp$idx_ok), ncol = info_par$par_num, byrow = TRUE)
-          if (any(model$m > 0L)) {
-            comp$par_tv[comp$idx_ok, ] <- comp$par_tv[comp$idx_ok, ] + sapply(1L:info_par$par_num, function(i) { comp$x[[i]][comp$idx_ok, , drop = FALSE] %*% comp$beta_list[[i]] })
-          }
-          cur_e <- rep(NA_real_, info_par$par_num)
-          for (j in comp$idx_ok) {
-            for (i in 1:info_par$par_num) {
-              cur_e[i] <- sum(comp$err_tv[j - seq_along(comp$phi_list[[i]]), i] * comp$phi_list[[i]]) + sum(comp$score_tv[j - seq_along(comp$alpha_list[[i]]), i] * comp$alpha_list[[i]])
-            }
-            comp$err_tv[j, ] <- cur_e
-            comp$par_tv[j, ] <- comp$par_tv[j, ] + cur_e
-            comp$y[j, ] <- fun$random(t = 1L, f = comp$par_tv[j, , drop = FALSE])
-            comp$score_tv[j, ] <- fun$score(y = comp$y[j, , drop = FALSE], f = comp$par_tv[j, , drop = FALSE])
-          }
-          comp$est_details$data$y <- comp$y[(comp$pre_num + comp$burn_num + 1L):comp$full_num, , drop = FALSE]
-          comp$result_optim <- do.call(control$optim_function, args = c(list(obj_fun = likelihood_objective, theta_start = comp$theta_start, theta_bound_lower = comp$theta_bound_lower, theta_bound_upper = comp$theta_bound_upper, est_details = comp$est_details), control$optim_arguments))
-          coef_set <- convert_theta_vector_to_coef_vector(comp$result_optim$theta_optim, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
-          return(coef_set)
-        }
+        comp$est_details$data$y <- y[(pre_num + burn_num + 1L):full_num, , drop = FALSE]
+        result_optim <- do.call(control$optim_function, args = c(list(obj_fun = likelihood_objective, theta_start = comp$theta_start, theta_bound_lower = comp$theta_bound_lower, theta_bound_upper = comp$theta_bound_upper, est_details = comp$est_details), control$optim_arguments))
+        coef_set <- convert_theta_vector_to_coef_vector(result_optim$theta_optim, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
+        return(coef_set)
       }
-      comp$coef_list <- do.call(control$parallel_function, args = c(list(run_num = comp$rep_boot, run_fun = comp$run_fun, run_details = comp$run_details), control$parallel_arguments))
-      bootstrap$coef_set <- matrix(unlist(comp$coef_list), nrow = comp$rep_boot, byrow = TRUE)
+    } else if (bootstrap$method == "parametric" && model$regress == "joint") {
+      comp$run_details <- list(data = data, model = model, control = control, info_distr = info_distr, info_par = info_par, info_coef = info_coef, comp = comp)
+      comp$run_fun <- function(id, run_details) {
+        data <- run_details$data
+        model <- run_details$model
+        control <- run_details$control
+        info_distr <- run_details$info_distr
+        info_par <- run_details$info_par
+        info_coef <- run_details$info_coef
+        comp <- run_details$comp
+        pre_num <- max(c(model$p, model$q, 1L))
+        burn_num <- 10L + max(model$p) + max(model$q)
+        full_num <- pre_num + burn_num + model$t
+        average_x <- lapply(1:info_par$par_num, function(i) { colMeans(data$x[[i]], na.rm = TRUE) })
+        y <- matrix(NA_real_, nrow = full_num, ncol = model$n)
+        x <- lapply(1:info_par$par_num, function(i) { rbind(matrix(NA_real_, nrow = pre_num, ncol = model$m[i]), matrix(average_x[[i]], nrow = burn_num, ncol = model$m[i], byrow = TRUE), data$x[[i]]) })
+        par_tv <- matrix(NA_real_, nrow = full_num, ncol = info_par$par_num)
+        score_tv <- matrix(NA_real_, nrow = full_num, ncol = info_par$par_num)
+        idx_na <- which(rowSums(sapply(x, function(e) { rowSums(is.na(e)) })) > 0L)
+        idx_ok <- which(!((1:full_num) %in% idx_na))
+        struc <- convert_coef_vector_to_struc_list(coef_vec = model$coef_est, m = model$m, p = model$p, q = model$q, par_names = info_par$par_names, par_of_coef_names = info_coef$par_of_coef_names)
+        omega_vector <- sapply(struc, function(e) { e$omega })
+        beta_list <- lapply(struc, function(e) { e$beta })
+        alpha_list <- lapply(struc, function(e) { e$alpha })
+        phi_list <- lapply(struc, function(e) { e$phi })
+        par_init <- model$par_init
+        if (any(is.na(par_init))) {
+          par_unc <- sapply(1:info_par$par_num, function(i) { (omega_vector[i] + average_x[[i]] %*% beta_list[[i]]) / (1 - sum(phi_list[[i]])) })
+          par_init[is.na(par_init)] <- par_unc[is.na(par_init)]
+        }
+        if (length(idx_na) > 0L) {
+          par_tv[idx_na, ] <- matrix(par_init, nrow = length(idx_na), ncol = info_par$par_num, byrow = TRUE)
+          score_tv[idx_na, ] <- 0
+        }
+        par_tv[idx_ok, ] <- matrix(omega_vector, nrow = length(idx_ok), ncol = info_par$par_num, byrow = TRUE)
+        if (any(model$m > 0L)) {
+          par_tv[idx_ok, ] <- par_tv[idx_ok, ] + sapply(1L:info_par$par_num, function(i) { x[[i]][idx_ok, , drop = FALSE] %*% beta_list[[i]] })
+        }
+        cur_e <- rep(NA_real_, info_par$par_num)
+        for (j in idx_ok) {
+          for (i in 1:info_par$par_num) {
+            cur_e[i] <- sum(par_tv[j - seq_along(phi_list[[i]]), i] * phi_list[[i]]) + sum(score_tv[j - seq_along(alpha_list[[i]]), i] * alpha_list[[i]])
+          }
+          par_tv[j, ] <- par_tv[j, ] + cur_e
+          y[j, ] <- fun$random(t = 1L, f = par_tv[j, , drop = FALSE])
+          score_tv[j, ] <- fun$score(y = y[j, , drop = FALSE], f = par_tv[j, , drop = FALSE])
+        }
+        comp$est_details$data$y <- y[(pre_num + burn_num + 1L):full_num, , drop = FALSE]
+        result_optim <- do.call(control$optim_function, args = c(list(obj_fun = likelihood_objective, theta_start = comp$theta_start, theta_bound_lower = comp$theta_bound_lower, theta_bound_upper = comp$theta_bound_upper, est_details = comp$est_details), control$optim_arguments))
+        coef_set <- convert_theta_vector_to_coef_vector(result_optim$theta_optim, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
+        return(coef_set)
+      }
+    } else if (bootstrap$method == "parametric" && model$regress == "sep") {
+      comp$run_details <- list(data = data, model = model, control = control, info_distr = info_distr, info_par = info_par, info_coef = info_coef, comp = comp)
+      comp$run_fun <- function(id, run_details) {
+        data <- run_details$data
+        model <- run_details$model
+        control <- run_details$control
+        info_distr <- run_details$info_distr
+        info_par <- run_details$info_par
+        info_coef <- run_details$info_coef
+        comp <- run_details$comp
+        pre_num <- max(c(model$p, model$q, 1L))
+        burn_num <- 10L + max(model$p) + max(model$q)
+        full_num <- pre_num + burn_num + model$t
+        average_x <- lapply(1:info_par$par_num, function(i) { colMeans(data$x[[i]], na.rm = TRUE) })
+        y <- matrix(NA_real_, nrow = full_num, ncol = model$n)
+        x <- lapply(1:info_par$par_num, function(i) { rbind(matrix(NA_real_, nrow = pre_num, ncol = model$m[i]), matrix(average_x[[i]], nrow = burn_num, ncol = model$m[i], byrow = TRUE), data$x[[i]]) })
+        par_tv <- matrix(NA_real_, nrow = full_num, ncol = info_par$par_num)
+        score_tv <- matrix(NA_real_, nrow = full_num, ncol = info_par$par_num)
+        idx_na <- which(rowSums(sapply(x, function(e) { rowSums(is.na(e)) })) > 0L)
+        idx_ok <- which(!((1:full_num) %in% idx_na))
+        struc <- convert_coef_vector_to_struc_list(coef_vec = model$coef_est, m = model$m, p = model$p, q = model$q, par_names = info_par$par_names, par_of_coef_names = info_coef$par_of_coef_names)
+        omega_vector <- sapply(struc, function(e) { e$omega })
+        beta_list <- lapply(struc, function(e) { e$beta })
+        alpha_list <- lapply(struc, function(e) { e$alpha })
+        phi_list <- lapply(struc, function(e) { e$phi })
+        err_tv <- matrix(NA_real_, nrow = full_num, ncol = info_par$par_num)
+        err_init <- model$par_init
+        par_init <- model$par_init
+        if (any(is.na(model$par_init))) {
+          par_unc <- sapply(1:info_par$par_num, function(i) { (omega_vector[i] + average_x[[i]] %*% beta_list[[i]]) })
+          par_init[is.na(par_init)] <- par_unc[is.na(par_init)]
+          err_init[is.na(err_init)] <- 0
+        }
+        if (length(idx_na) > 0L) {
+          err_tv[idx_na, ] <- matrix(err_init, nrow = length(idx_na), ncol = info_par$par_num, byrow = TRUE)
+          par_tv[idx_na, ] <- matrix(par_init, nrow = length(idx_na), ncol = info_par$par_num, byrow = TRUE)
+          score_tv[idx_na, ] <- 0
+        }
+        par_tv[idx_ok, ] <- matrix(omega_vector, nrow = length(idx_ok), ncol = info_par$par_num, byrow = TRUE)
+        if (any(model$m > 0L)) {
+          par_tv[idx_ok, ] <- par_tv[idx_ok, ] + sapply(1L:info_par$par_num, function(i) { x[[i]][idx_ok, , drop = FALSE] %*% beta_list[[i]] })
+        }
+        cur_e <- rep(NA_real_, info_par$par_num)
+        for (j in idx_ok) {
+          for (i in 1:info_par$par_num) {
+            cur_e[i] <- sum(err_tv[j - seq_along(phi_list[[i]]), i] * phi_list[[i]]) + sum(score_tv[j - seq_along(alpha_list[[i]]), i] * alpha_list[[i]])
+          }
+          err_tv[j, ] <- cur_e
+          par_tv[j, ] <- par_tv[j, ] + cur_e
+          y[j, ] <- fun$random(t = 1L, f = par_tv[j, , drop = FALSE])
+          score_tv[j, ] <- fun$score(y = y[j, , drop = FALSE], f = par_tv[j, , drop = FALSE])
+        }
+        comp$est_details$data$y <- y[(pre_num + burn_num + 1L):full_num, , drop = FALSE]
+        result_optim <- do.call(control$optim_function, args = c(list(obj_fun = likelihood_objective, theta_start = comp$theta_start, theta_bound_lower = comp$theta_bound_lower, theta_bound_upper = comp$theta_bound_upper, est_details = comp$est_details), control$optim_arguments))
+        coef_set <- convert_theta_vector_to_coef_vector(result_optim$theta_optim, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
+        return(coef_set)
+      }
+    } else if (bootstrap$method == "simple_block") {
+      bootstrap$block_length <- check_my_block_length(block_length = block_length, t = model$t)
+      comp$run_details <- list(data = data, model = model, control = control, bootstrap = bootstrap, comp = comp)
+      comp$run_fun <- function(id, run_details) {
+        data <- run_details$data
+        model <- run_details$model
+        control <- run_details$control
+        bootstrap <- run_details$bootstrap
+        comp <- run_details$comp
+        t <- model$t
+        block_length <- bootstrap$block_length
+        circle <- c(1:t, 1:t)
+        block_num <- round(t / block_length)
+        block_range <- rep(block_length, block_num)
+        if (block_length * block_num > t) {
+          block_range[sample(1:block_num, size = block_length * block_num - t, replace = FALSE)] <- block_length - 1
+        } else if (block_length * block_num < t) {
+          block_range[sample(1:block_num, size = t - block_length * block_num, replace = FALSE)] <- block_length + 1
+        }
+        block_start <- (sample(1:block_num, size = block_num, replace = TRUE) - 1) * block_length + 1
+        block_end <- block_start + block_range - 1
+        idx <- circle[unlist(lapply(1:block_num, FUN = function(i) { block_start[i]:block_end[i] }))]
+        comp$est_details$data$y <- data$y[idx, , drop = FALSE]
+        comp$est_details$data$x <- lapply(data$x, function(xx) { xx[idx, , drop = FALSE] })
+        result_optim <- do.call(control$optim_function, args = c(list(obj_fun = likelihood_objective, theta_start = comp$theta_start, theta_bound_lower = comp$theta_bound_lower, theta_bound_upper = comp$theta_bound_upper, est_details = comp$est_details), control$optim_arguments))
+        coef_set <- convert_theta_vector_to_coef_vector(result_optim$theta_optim, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
+        return(coef_set)
+      }
+    } else if (bootstrap$method == "moving_block") {
+      bootstrap$block_length <- check_my_block_length(block_length = block_length, t = model$t)
+      comp$run_details <- list(data = data, model = model, control = control, bootstrap = bootstrap, comp = comp)
+      comp$run_fun <- function(id, run_details) {
+        data <- run_details$data
+        model <- run_details$model
+        control <- run_details$control
+        bootstrap <- run_details$bootstrap
+        comp <- run_details$comp
+        t <- model$t
+        block_length <- bootstrap$block_length
+        circle <- c(1:t, 1:t)
+        block_num <- round(t / block_length)
+        block_range <- rep(block_length, block_num)
+        if (block_length * block_num > t) {
+          block_range[sample(1:block_num, size = block_length * block_num - t, replace = FALSE)] <- block_length - 1
+        } else if (block_length * block_num < t) {
+          block_range[sample(1:block_num, size = t - block_length * block_num, replace = FALSE)] <- block_length + 1
+        }
+        block_start <- sample(1:t, size = block_num, replace = TRUE)
+        block_end <- block_start + block_range - 1
+        idx <- circle[unlist(lapply(1:block_num, FUN = function(i) { block_start[i]:block_end[i] }))]
+        comp$est_details$data$y <- data$y[idx, , drop = FALSE]
+        comp$est_details$data$x <- lapply(data$x, function(xx) { xx[idx, , drop = FALSE] })
+        result_optim <- do.call(control$optim_function, args = c(list(obj_fun = likelihood_objective, theta_start = comp$theta_start, theta_bound_lower = comp$theta_bound_lower, theta_bound_upper = comp$theta_bound_upper, est_details = comp$est_details), control$optim_arguments))
+        coef_set <- convert_theta_vector_to_coef_vector(result_optim$theta_optim, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
+        return(coef_set)
+      }
+    } else if (bootstrap$method == "stationary_block") {
+      bootstrap$block_length <- check_my_block_length(block_length = block_length, t = model$t)
+      comp$run_details <- list(data = data, model = model, control = control, bootstrap = bootstrap, comp = comp)
+      comp$run_fun <- function(id, run_details) {
+        data <- run_details$data
+        model <- run_details$model
+        control <- run_details$control
+        bootstrap <- run_details$bootstrap
+        comp <- run_details$comp
+        t <- model$t
+        block_length <- bootstrap$block_length
+        circle <- c(1:t, 1:t)
+        block_prob <- 1 / block_length
+        block_range <- diff(which(c(TRUE, sample(c(TRUE, FALSE), size = t - 1, replace = TRUE, prob = c(block_prob, 1 - block_prob)), TRUE)))
+        block_num <- length(block_range)
+        block_start <- sample(1:t, size = block_num, replace = TRUE)
+        block_end <- block_start + block_range - 1
+        idx <- circle[unlist(lapply(1:block_num, FUN = function(i) { block_start[i]:block_end[i] }))]
+        comp$est_details$data$y <- data$y[idx, , drop = FALSE]
+        comp$est_details$data$x <- lapply(data$x, function(xx) { xx[idx, , drop = FALSE] })
+        result_optim <- do.call(control$optim_function, args = c(list(obj_fun = likelihood_objective, theta_start = comp$theta_start, theta_bound_lower = comp$theta_bound_lower, theta_bound_upper = comp$theta_bound_upper, est_details = comp$est_details), control$optim_arguments))
+        coef_set <- convert_theta_vector_to_coef_vector(result_optim$theta_optim, coef_fix_value = model$coef_fix_value, coef_fix_other = model$coef_fix_other)
+        return(coef_set)
+      }
     }
+    comp$coef_list <- do.call(control$parallel_function, args = c(list(run_num = comp$rep_boot, run_fun = comp$run_fun, run_details = comp$run_details), control$parallel_arguments))
+    bootstrap$coef_set <- name_matrix(matrix(unlist(comp$coef_list), nrow = comp$rep_boot, byrow = TRUE), paste0("s", 1:comp$rep_boot), info_coef$coef_names, drop = c(FALSE, FALSE))
     info_data <- info_data(y = data$y, x = data$x)
     data$y <- name_matrix(data$y, info_data$index_time, info_data$index_series, drop = c(FALSE, TRUE))
     data$x <- name_list_of_matrices(data$x, info_par$par_names, info_data$index_time_list, info_data$index_vars_list, drop = c(FALSE, TRUE), zero = c(FALSE, TRUE))
-    bootstrap$coef_set <- name_matrix(bootstrap$coef_set, paste0("s", 1:comp$rep_boot), info_coef$coef_names, drop = c(FALSE, FALSE))
     bootstrap$coef_mean <- name_vector(colMeans(bootstrap$coef_set), info_coef$coef_names)
     bootstrap$coef_vcov <- name_matrix(stats::cov(bootstrap$coef_set), info_coef$coef_names, info_coef$coef_names,  drop = c(FALSE, FALSE))
     bootstrap$coef_sd <-  name_vector(sqrt(diag(bootstrap$coef_vcov)), info_coef$coef_names)